// test/integration/test/ports/prisma/functional/issues-23902/_fixture/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly viewCount: CodecTypes['pg/int4@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly viewCount: CodecTypes['pg/int4@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['output'] | null;
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly viewCount: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['input'] | null;
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly viewCount: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;